requirements.txt
Keras==2.2.4
matplotlib==3.0.2
numpy==1.15.4
scipy==1.1.0
tensorflow-gpu==1.12.0
tqdm==4.28.1
opencv-python==3.4.4.19
Pillow==5.3.0
requests==2.21.0
